Add benchmark infrastructure
This commit adds a Docker container for the benchmark client.
The client is based on a Node JS image and compiles wrk2.
A Node application drives wrk2 and runs benchmark plans; results
are then uploaded as GitHub gists and notifies the operator
through a private Slack Channel when done.

This container is uploaded to GCP.

This commit also ships with the scripts that start/stop benchmarks.
Using the bechmark script, we can copy executable JARs on a
private GCP storage bucket. Once there, we can start a benchmark
and the script does the following:

1. Create a server instance, copy the benchmark app from the bucket
and run it
2. Wait until the server app is up
3. Create a client instance with the docker image and run the
benchmark against the internal IP of the server
4. Once the results are recorded, delete all service instances